+++ src/ecs/World.hpp
@@ -1,0 +1,13 @@
+/*
+** EPITECH PROJECT, 2022
+** Bomberman: Global Offensive
+** File description:
+** ECS - World
+*/
+
+#ifndef ECS_WORLD_HPP_
+#define ECS_WORLD_HPP_
+
+#include "ecs/world/World.hpp"
+
+#endif // !defined(ECS_WORLD_HPP_)
