+++ src/ecs/resource/CMakeLists.txt
@@ -1,0 +1,10 @@
+set(INCROOT ${INCROOT}/resource)
+set(SRCROOT ${SRCROOT}/resource)
+
+set(SRC
+	${SRC}
+    ${INCROOT}/Entities.cpp
+    ${INCROOT}/Entities.hpp
+    ${INCROOT}/Resource.hpp
+	PARENT_SCOPE
+)
