+++ src/ecs/system/CMakeLists.txt
@@ -1,0 +1,9 @@
+set(INCROOT ${INCROOT}/system)
+set(SRCROOT ${SRCROOT}/system)
+
+set(SRC
+	${SRC}
+    ${INCROOT}/System.hpp
+    ${INCROOT}/SystemData.hpp
+	PARENT_SCOPE
+)
