+++ src/ecs/world/CMakeLists.txt
@@ -3,9 +3,7 @@ set(SRCROOT ${SRCROOT}/world)
 
 set(SRC
 	${SRC}
-    ${INCROOT}/Entities.cpp
-    ${INCROOT}/Entities.hpp
-    ${INCROOT}/Resource.hpp
+    ${INCROOT}/Instances.hpp
     ${SRCROOT}/World.cpp
     ${INCROOT}/World.hpp
 	PARENT_SCOPE
